Move legend-item into separate component
